refact: QuantitySelector를 공통 컴포넌트로 추출
